Move syntax highlighting into a trait
