+++ src/allejo/stakx/MarkupEngine/MarkdownEngine.php
@@ -15,7 +15,7 @@ use function HighlightUtilities\splitCodeIntoArray;
 
 class MarkdownEngine extends \ParsedownExtra implements MarkupEngineInterface
 {
-    protected $highlighter;
+    use SyntaxHighlighterTrait;
 
     public function __construct()
     {
@@ -67,94 +67,12 @@ class MarkdownEngine extends \ParsedownExtra implements MarkupEngineInterface
         if (isset($block['element']['text']['attributes']) && Service::hasRunTimeFlag(RuntimeStatus::USING_HIGHLIGHTER))
         {
             $cssClass = $block['element']['text']['attributes']['class'];
-            $langDef = $this->parseInfoString($cssClass);
+            $block['markup'] = $this->highlightCode($cssClass, $block['element']['text']['text']);
 
-            try
-            {
-                $highlighted = $this->highlighter->highlight($langDef['language'], $block['element']['text']['text']);
-                $value = $highlighted->value;
-
-                if (Service::hasRunTimeFlag(RuntimeStatus::USING_LINE_NUMBERS) || count($langDef['selectedLines']) > 0)
-                {
-                    $lines = splitCodeIntoArray($value);
-                    $value = '';
-
-                    foreach ($lines as $i => $line)
-                    {
-                        // `$i + 1` since our line numbers are indexed starting at 1
-                        $value .= vsprintf("<div class=\"loc%s\"><span>%s</span></div>\n", [
-                            isset($langDef['selectedLines'][$i + 1]) ? ' highlighted' : '',
-                            $line,
-                        ]);
-                    }
-                }
-
-                $block['markup'] = vsprintf('<pre><code class="hljs language-%s">%s</code></pre>', [
-                    $langDef['language'],
-                    $value,
-                ]);
-
-                // Only return the block if Highlighter knew the language and how to handle it.
-                return $block;
-            }
-            // Exception thrown when language not supported
-            catch (\DomainException $exception)
-            {
-                trigger_error("An unsupported language (${langDef['language']}) was detected in a code block", E_USER_WARNING);
-            }
-            catch (\Exception $e)
-            {
-                trigger_error('An error has occurred in the highlight.php language definition files', E_USER_WARNING);
-            }
+            return $block;
         }
 
         return parent::blockFencedCodeComplete($block);
-    }
-
-    private function parseInfoString($infoString)
-    {
-        $infoString = substr($infoString, 9);
-        $definition = [
-            'language' => $infoString,
-            'selectedLines' => [],
-        ];
-
-        $bracePos = strpos($infoString, '{');
-
-        if ($bracePos === false)
-        {
-            return $definition;
-        }
-
-        $definition['language'] = substr($infoString, 0, $bracePos);
-        $lineDefinition = substr($infoString, $bracePos + 1, -1);
-        $lineNumbers = explode(',', $lineDefinition);
-
-        foreach ($lineNumbers as $lineNumber)
-        {
-            if (strpos($lineNumber, '-') === false)
-            {
-                $definition['selectedLines'][intval($lineNumber)] = true;
-                continue;
-            }
-
-            $extremes = explode('-', $lineNumber);
-
-            if (count($extremes) !== 2)
-            {
-                continue;
-            }
-
-            $start = intval($extremes[0]);
-            $end = intval($extremes[1]);
-
-            for ($i = $start; $i <= $end; $i++)
-            {
-                $definition['selectedLines'][$i] = true;
-            }
-        }
-
-        return $definition;
     }
 
     ///
